+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## Fixes
+
+-   Keep redirectToPath's query params while redirecting to it from auth
+
 ## [0.31.1] - 2023-02-12
 
 ### Changes